RavenDB-17793 : block prefix sharding usage, use it only in tests
