+++ src/Raven.Server/ServerWide/ServerStore.cs
@@ -2783,6 +2783,9 @@ namespace Raven.Server.ServerWide
             }
             else
             {
+                if (BlockPrefixedSharding && record.Sharding.Prefixed is { Count: > 0 })
+                    throw new InvalidOperationException("Cannot use prefixed sharding, this feature is currently blocked");
+
                 InitializeTopology(record.Sharding.Orchestrator.Topology);
 
                 foreach (var (shardNumber, shardTopology) in record.Sharding.Shards)
