+++ test/SlowTests/Sharding/Backup/ShardedRestoreBackupTests.cs
@@ -23,7 +23,6 @@ using Raven.Server.Utils;
 using Raven.Tests.Core.Utils.Entities;
 using SlowTests.Server.Documents.PeriodicBackup;
 using Tests.Infrastructure;
-using Tests.Infrastructure.Entities;
 using Xunit;
 using Xunit.Abstractions;
 using BackupConfiguration = Raven.Server.Config.Categories.BackupConfiguration;
@@ -886,168 +885,6 @@ namespace SlowTests.Sharding.Backup
                             var doc = await session.LoadAsync<User>($"users/{i}/$b");
                             Assert.NotNull(doc);
                         }
-                    }
-                }
-            }
-        }
-
-        [RavenFact(RavenTestCategory.BackupExportImport | RavenTestCategory.Sharding)]
-        public async Task BackupAndRestoreShardedDatabase_ShouldPreservePrefixedSettingsAndBucketRanges()
-        {
-            using (var store = Sharding.GetDocumentStore())
-            {
-                var databaseRecord = await store.Maintenance.Server.SendAsync(new GetDatabaseRecordOperation(store.Database));
-                databaseRecord.Sharding.Prefixed ??= new List<PrefixedShardingSetting>();
-                databaseRecord.Sharding.Prefixed.Add(new PrefixedShardingSetting
-                {
-                    Prefix = "users/",
-                    Shards = new List<int> { 0 }
-                });
-                databaseRecord.Sharding.Prefixed.Add(new PrefixedShardingSetting
-                {
-                    Prefix = "orders/",
-                    Shards = new List<int> { 1 }
-                });
-                databaseRecord.Sharding.Prefixed.Add(new PrefixedShardingSetting
-                {
-                    Prefix = "employees/", Shards = new List<int> { 2 }
-                });
-
-                await store.Maintenance.Server.SendAsync(new UpdateDatabaseOperation(databaseRecord, replicationFactor: 1, databaseRecord.Etag));
-                var sharding = await Sharding.GetShardingConfigurationAsync(store);
-
-                Assert.Equal(6, sharding.BucketRanges.Count);
-                Assert.Equal(ShardHelper.NumberOfBuckets, sharding.BucketRanges[3].BucketRangeStart);
-                Assert.Equal(ShardHelper.NumberOfBuckets * 2, sharding.BucketRanges[4].BucketRangeStart);
-                Assert.Equal(ShardHelper.NumberOfBuckets * 3, sharding.BucketRanges[5].BucketRangeStart);
-
-                using (var session = store.OpenAsyncSession())
-                {
-                    for (int i = 0; i < 10; i++)
-                    {
-                        await session.StoreAsync(new User(), $"users/{i}");
-                        await session.StoreAsync(new Order(), $"orders/{i}");
-                        await session.StoreAsync(new Employee(), $"employees/{i}");
-                    }
-
-                    await session.SaveChangesAsync();
-                }
-
-                using (var session = store.OpenAsyncSession(database: ShardHelper.ToShardName(store.Database, 0)))
-                {
-                    for (int i = 0; i < 10; i++)
-                    {
-                        var doc = await session.LoadAsync<User>($"users/{i}");
-                        Assert.NotNull(doc);
-                    }
-                }
-                using (var session = store.OpenAsyncSession(database: ShardHelper.ToShardName(store.Database, 1)))
-                {
-                    for (int i = 0; i < 10; i++)
-                    {
-                        var doc = await session.LoadAsync<Order>($"orders/{i}");
-                        Assert.NotNull(doc);
-                    }
-                }
-                using (var session = store.OpenAsyncSession(database: ShardHelper.ToShardName(store.Database, 2)))
-                {
-                    for (int i = 0; i < 10; i++)
-                    {
-                        var doc = await session.LoadAsync<Employee>($"employees/{i}");
-                        Assert.NotNull(doc);
-                    }
-                }
-
-                var waitHandles = await Sharding.Backup.WaitForBackupToComplete(store);
-                var backupPath = NewDataPath(suffix: "BackupFolder");
-                var config = Backup.CreateBackupConfiguration(backupPath);
-
-                await Sharding.Backup.UpdateConfigurationAndRunBackupAsync(Server, store, config);
-                Assert.True(WaitHandle.WaitAll(waitHandles, TimeSpan.FromMinutes(1)));
-
-                var dirs = Directory.GetDirectories(backupPath);
-                Assert.Equal(3, dirs.Length);
-
-                sharding = await Sharding.GetShardingConfigurationAsync(store);
-                var settings = Sharding.Backup.GenerateShardRestoreSettings(dirs, sharding);
-
-                // restore the database with a different name
-                var restoredDatabaseName = $"restored_database-{Guid.NewGuid()}";
-                using (Sharding.Backup.ReadOnly(backupPath))
-                using (Backup.RestoreDatabase(store, new RestoreBackupConfiguration
-                {
-                    DatabaseName = restoredDatabaseName,
-                    ShardRestoreSettings = settings
-
-                }, timeout: TimeSpan.FromSeconds(60)))
-                {
-                    var newDatabaseRecord = await store.Maintenance.Server.SendAsync(new GetDatabaseRecordOperation(restoredDatabaseName));
-                    Assert.Equal(3, newDatabaseRecord.Sharding.Shards.Count);
-                    Assert.Equal(3, newDatabaseRecord.Sharding.Prefixed.Count);
-
-                    var usersPrefixSetting = newDatabaseRecord.Sharding.Prefixed[0];
-                    Assert.Equal("users/", usersPrefixSetting.Prefix);
-                    Assert.Equal(1, usersPrefixSetting.Shards.Count);
-                    Assert.Equal(0, usersPrefixSetting.Shards[0]);
-                    Assert.Equal(ShardHelper.NumberOfBuckets, usersPrefixSetting.BucketRangeStart);
-
-                    var ordersPrefixSetting = newDatabaseRecord.Sharding.Prefixed[1];
-                    Assert.Equal("orders/", ordersPrefixSetting.Prefix);
-                    Assert.Equal(1, ordersPrefixSetting.Shards.Count);
-                    Assert.Equal(1, ordersPrefixSetting.Shards[0]);
-                    Assert.Equal(ShardHelper.NumberOfBuckets * 2, ordersPrefixSetting.BucketRangeStart);
-
-                    var employeesPrefixSetting = newDatabaseRecord.Sharding.Prefixed[2];
-                    Assert.Equal("employees/", employeesPrefixSetting.Prefix);
-                    Assert.Equal(1, employeesPrefixSetting.Shards.Count);
-                    Assert.Equal(2, employeesPrefixSetting.Shards[0]);
-                    Assert.Equal(ShardHelper.NumberOfBuckets * 3, employeesPrefixSetting.BucketRangeStart);
-
-                    Assert.Equal(6, sharding.BucketRanges.Count);
-                    Assert.Equal(ShardHelper.NumberOfBuckets, sharding.BucketRanges[3].BucketRangeStart);
-                    Assert.Equal(ShardHelper.NumberOfBuckets * 2, sharding.BucketRanges[4].BucketRangeStart);
-                    Assert.Equal(ShardHelper.NumberOfBuckets * 3, sharding.BucketRanges[5].BucketRangeStart);
-
-                    using (var session = store.OpenAsyncSession(database: restoredDatabaseName))
-                    {
-                        for (int i = 0; i < 10; i++)
-                        {
-                            var user = await session.LoadAsync<User>($"users/{i}");
-                            Assert.NotNull(user);
-
-                            var order = await session.LoadAsync<Order>($"orders/{i}");
-                            Assert.NotNull(order);
-
-                            var employee = await session.LoadAsync<Employee>($"employees/{i}");
-                            Assert.NotNull(employee);
-                        }
-                    }
-
-                    using (var session = store.OpenAsyncSession(database: restoredDatabaseName))
-                    {
-                        await session.StoreAsync(new User(), "users/11");
-                        await session.StoreAsync(new Order(), "orders/11");
-                        await session.StoreAsync(new Employee(), "employees/11");
-
-                        await session.SaveChangesAsync();
-                    }
-
-                    using (var session = store.OpenAsyncSession(database: ShardHelper.ToShardName(restoredDatabaseName, 0)))
-                    {
-                        var user = await session.LoadAsync<User>("users/11");
-                        Assert.NotNull(user);
-                    }
-
-                    using (var session = store.OpenAsyncSession(database: ShardHelper.ToShardName(restoredDatabaseName, 1)))
-                    {
-                        var order = await session.LoadAsync<Order>("orders/11");
-                        Assert.NotNull(order);
-                    }
-
-                    using (var session = store.OpenAsyncSession(database: ShardHelper.ToShardName(restoredDatabaseName, 2)))
-                    {
-                        var employee = await session.LoadAsync<Employee>("employees/11");
-                        Assert.NotNull(employee);
                     }
                 }
             }
